requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
python-dotenv==1.0.1
openai==1.61.1
httpx==0.28.1
gspread==6.2.1
requests==2.32.3
twilio==9.6.1
supabase==2.13.0
pinecone-client==5.0.1
google-auth==2.40.2
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.170.0 